backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0